# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_transport_http_v1.getUserResponse:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      email:
        type: string
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API for Mile-Do, simple clone of TickTick.
  termsOfService: http://swagger.io/terms/
  title: Mile-Do API
  version: "1.0"
paths:
...
  /users/:
    delete:
      consumes:
      - application/json
      description: logout from user account
      produces:
      - application/json
      responses:
        "200":
          description: successful log out
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: logout user
      tags:
      - auth
  /users/me:
    get:
      consumes:
      - application/json
      description: get user account by bearer token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.getUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get user info
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"